re-arrange-component
