+++ Dockerfile
@@ -1,3 +1,28 @@
+# # Stage 1 - Build the base
+# FROM node:18-alpine AS base
+# WORKDIR /app
+# COPY src ./src
+# COPY package*.json ./
+# COPY tsconfig*.json ./
+# COPY prisma ./prisma
+# RUN npm install
+
+# # Stage 2 - Build the app
+# FROM base AS build
+# WORKDIR /app
+# RUN npm run build
+
+# # Stage 3 - Production
+# FROM node:18-alpine
+# WORKDIR /app
+# COPY package*.json ./
+# RUN npm install --only-production
+# COPY --from=build /app/build ./
+# COPY prisma ./prisma
+# RUN npx prisma generate
+
+# CMD ["node", "main.js"]
+
 # Stage 1 - Build the base
 FROM node:18-alpine AS base
 WORKDIR /app
@@ -16,9 +41,12 @@ RUN npm run build
 FROM node:18-alpine
 WORKDIR /app
 COPY package*.json ./
-RUN npm install --only-production
-COPY --from=build /app/build ./
+RUN npm install --only=production
+COPY --from=build /app/build ./src
 COPY prisma ./prisma
 RUN npx prisma generate
 
-CMD ["node", "main.js"]
+# Update CMD to match entry point
+CMD ["node", "src/server.js"]
+
+
